Prepare for BoringSSL update.
This change tweaks things as needed so that the code will compile
against both the BoringSSL that's currently in Android and a version
from upstream. The BORINGSSL_201509 define is temporary to allow the
switch to happen without breaking the build and a followup change will
remove it.

In this case, we backed off on removing |ERR_print_errors|, despite it
being a wart in the layering of the code.

(cherry picked from commit 7a688f6e8774920cb8cfc52bf2e3c16976e8ff23)

Change-Id: I2f2f567cfc37fe2d5a69742302c5a92d289b814f